https://cae-jira.jpl.nasa.gov/browse/IMCEI-173
The CMOF constraints for associations already force
such associations to be binary but fall short of forcing
them to be logically or semantically directed. For OTI,
CMOF associations are restricted to be logically or semantically
directed, which means that the order of association member ends
is either redundant with the logical or semantic ordering
or inconsistent with it. For code generation purposes,
preference is given to the logical or semantic ordering and in the
few cases where this is insufficient, the ordering follows
that of associaiton member ends.